L1TF/Linux: Fix "running an hypervisor" detection via sysfs
If we can look in `/sys/devices/system/cpu/vulnerabilities/l1tf`, and
we find 'VMX' in it, it means we can run KVM guests, and hence we are
running as an hypervisor.

Right now, on a system with VMX enabled, with KVM guests running, and
booted with "l1tf=off", I see this:
`
  CVE-2018-3646 aka 'Foreshadow-NG (VMM), L1 terminal fault'
  * Information from the /sys interface: VMX: vulnerable
  * This system is a host running a hypervisor:  NO
  ... ... ...
    * L1D flush enabled:  NO
  > STATUS:  NOT VULNERABLE  (this system is not running a hypervisor)
`
Which is wrong.

In fact, as said, guests are running, so we are an hypervisor,
and with "l1tf=off" (and hence L1D flush disabled) we are vulnerable.